refactor: updated auth options
